Restructored sender/receiver TCP connection and added vcmtpRecvv3::Stop().
Eliminated reconnection attempt of TCP connection when it breaks because the
decision on what to do needs to be made by the application.

Moved initialization of TCP receiver from vcmtpRecvv3::Start() to constructors
because a failure of the TCP receiver to initialize should be determined as
soon as possible.

Added vcmtpRecvv3::Stop() because using the destructor to stop a VCMTP receiver
results in the destructor being called twice (once for stopping and one at end-
of-life) and consequent memory corruption.

Added some thrown exceptions and documentation.

Added some commented-out debugging statements.